Fix error by race detector
